refactor(db): rename runner.ts to runner-job-queue.ts
Renamed schema file to reflect its actual content (only contains
runnerJobQueue table, runners table was removed in migration cleanup).
Updated all imports accordingly.
