+++ turbo/apps/web/app/api/runners/jobs/[id]/claim/route.ts
@@ -10,7 +10,7 @@ import {
 } from "@vm0/core";
 import { initServices } from "../../../../../../src/lib/init-services";
 import { agentRuns } from "../../../../../../src/db/schema/agent-run";
-import { runnerJobQueue } from "../../../../../../src/db/schema/runner";
+import { runnerJobQueue } from "../../../../../../src/db/schema/runner-job-queue";
 import { eq, and, isNull } from "drizzle-orm";
 import { getUserId } from "../../../../../../src/lib/auth/get-user-id";
 import { generateSandboxToken } from "../../../../../../src/lib/auth/sandbox-token";
